requirements.txt
streamlit
langchain
langchain-openai
langchain-community
python-dotenv